pyproject.toml
[project]
name = "semantic-catalogue"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "cjber", email = "cjberragan@gmail.com" }
]
dependencies = [
    "dagster>=1.7.8",
    "dagster-webserver>=1.7.8",
    "dagster-postgres>=0.23.15",
    "dagster-openai>=0.23.9",
    "requests>=2.31.0",
    "polars>=0.20.25",
    "python-dotenv>=1.0.1",
    "sickle>=0.7.0",
    "lxml>=5.2.2",
    "pydantic-settings>=2.3.4",
    "pdfminer-six>=20240706",
    "dateparser>=1.2.0",
    "langchain>=0.2.12",
    "langchain-community>=0.2.10",
    "langchain-pinecone>=0.1.3",
    "langchain-openai>=0.1.20",
    "langchain-experimental>=0.0.63",
    "langgraph>=0.1.19",
    "langchainhub>=0.1.20",
    "fastapi[standard]>=0.112.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.24.0",
    "sourcery-cli>=1.18.0",
    "ipdb>=0.13.13",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/semantic_catalogue"]

[tool.dagster]
module_name = "src.datastore"
